requirements.txt
certifi==2026.1.4
charset-normalizer==3.4.4
contourpy==1.3.3
cycler==0.12.1
filelock==3.20.3
fonttools==4.61.1
fsspec==2026.1.0
greenlet==3.3.0
idna==3.11
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
networkx==3.6.1
numpy==2.4.1
opencv-python==4.13.0.90
packaging==26.0
pandas==2.3.3
pillow==12.1.0
polars==1.37.1
polars-runtime-32==1.37.1
psutil==7.2.1
psycopg2-binary==2.9.11
pyaes==1.6.1
pyasn1==0.6.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
PyYAML==6.0.3
requests==2.32.5
rsa==4.9.1
scipy==1.17.0
six==1.17.0
SQLAlchemy==2.0.45
sympy==1.14.0
Telethon==1.42.0
torch==2.9.1
torchvision==0.24.1
typing_extensions==4.15.0
tzdata==2025.3
ultralytics==8.4.7
ultralytics-thop==2.0.18
urllib3==2.6.3
